Relegate fuse setup to hwmodel
Per comments from reviewers, this commit both adds comments regarding
fuse setup and also relegates all digest work and fuse setup to the
hwmodel, as it requires it but actual hardware should not.

The swsim, verilator, and FPGA do not have true OTP fuses so they need
to have OTP fuse data programmed on startup by the host SOC. In an actual
device, the OTP fuse data would be programmed during manufacturing.

As such all fuse handling has been moved out of the generic/main.c and
out of the Caliptra C API proper.
